Fixed Polygon Issue
Corrected issue with the order of the polygon points--used cross product of CL segment vector with respective left and right transect vectors to determine which side transect end point was on.

Also removed temporary code to read in csv export of intermediate shape files.

In addition, added some notes about CRS and units.